Minor changes to Travis config
Just like in Sequel, use a special .travis.gemfile for Travis,
don't have a general Gemfile.  Don't have a lock file for the
gemfile, it's not necessary.  Use the same rake/rspec version
constraints that Sequel uses for testing.

Only test on rbx-2, not both rbx-2.1 and rbx-2.2.  I don't think
it makes sense to test two versions of rbx.  Allow rbx to fail,
but not any of the other ruby versions to, since rbx on Travis
often fails for reasons unrelated to the code being tested.

Remove banners for Travis and Code Climate from the README. In
general if they are passing it means nothing, and failures are
often unrelated to the code being tested.  That's true for Travis
in my experience.  For Code Climate, I just don't see the point.