Deal with new rake and old rdoc in Rakefile
Newer versions of rake have a rake/rdoctask that raises a
RuntimeError if you try to load it.  This is a poor design
decision (raising a LoadError would be better), but that's life.

In any case, if rdoc/task doesn't exist and rake/rdoctask raises
an error, don't create an rdoc task, and have the rake actually
run instead of exiting with an error.